+++ pytsite/http/__init__.py
@@ -1,6 +1,11 @@
-# Public API
-from . import _error as error, _request as request, _response as response, _session as session
-
+"""PytSite HTTP Package
+"""
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+# Public API
+from . import _request as request, _response as response, _session as session, _error as error
+from ._request import Request
+from ._response import Response, Redirect as RedirectResponse, JSON as JSONResponse
+from ._session import Session